backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
openai>=1.44.0
pyyaml==6.0.1
SQLAlchemy==2.0.30
python-multipart==0.0.9
